@@ -45,7 +45,10 @@
     "webpack-dev-server": "2.3.0"
   },
   "dependencies": {
-    "react": "15.4.2",
-    "react-dom": "15.4.2"
+    "express": "4.16.4",
+    "react": "16.7.0-alpha.0",
+    "react-dom": "16.7.0-alpha.0",
+    "uuid": "^3.3.2",
+    "ws": "6.2.0"
   }
 }
